refactor(agent-adapters): split AgentEventAdapters into per-agent files
Mechanically split the 2872-LOC AgentEventAdapters.swift into focused
per-agent files under EventAdapters/ (one extension AgentEventBridge per
agent), plus a shared-helpers file. No behavior change: every static
adapter func keeps its signature and body verbatim.

The only cross-file coupling was droidTodoProgress(toolInput:) (used by
both droid and cursor); it and its DroidTodoProgressSnapshot move to the
shared file and widen from private to internal. Codex-notify is peeled
into its own file to keep each file under ~500 LOC.
